+++ modules/azure/aks/variables.tf
@@ -40,6 +40,9 @@ variable "aks_config" {
     version = string
     # Enables paid SKU for AKS and makes the default node pool HA
     production_grade = bool
+    # Will replace the default cluster auto scaler expander with a priority expander, 
+    # see https://github.com/kubernetes/autoscaler/blob/master/cluster-autoscaler/expander/priority/readme.md#configuration
+    priority_expander_config = optional(map(list(string)))
     node_pools = list(object({
       name           = string
       version        = string
@@ -165,7 +168,7 @@ variable "azure_metrics_identity" {
 variable "aks_audit_log_retention" {
   description = "The aks audit log retention in days, 0 = infinite"
   type        = number
-  default     = 180
+  default     = 365
 }
 
 variable "log_eventhub_name" {