+++ modules/azure/core/main.tf
@@ -2,6 +2,8 @@
   * # Core
   *
   * This module is used to create core resources like virtual network for the subscription.
+  * This module assumes that you have a RG called `rg-<env>-<location_short>-log`.
+  * Easiest is to define this RG in the governance module.
   */
 
 terraform {
@@ -9,7 +11,7 @@ terraform {
 
   required_providers {
     azurerm = {
-      version = "3.28.0"
+      version = "3.38.0"
       source  = "hashicorp/azurerm"
     }
     azuread = {