+++ modules/kubernetes/node-ttl/variables.tf
@@ -1,0 +1,4 @@
+variable "status_config_map_namespace" {
+  description = "Namespace where Cluster Autoscaler status ConfigMap is created"
+  type        = string
+}